% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cornejo2025}
\alias{Cornejo2025}
\title{Cornejo2025 longitudinal dataset measured in transgender persons}
\format{
\subsection{'Cornejo2025'}{

A list object with three elements:
\describe{
\item{Tongue}{List object of the tongue longitudinal microbiota data.}
\item{Saliva}{List object of the saliva longitudinal microbiota data.}
\item{Cytokines}{List object of the longitudinal salivary cytokines data.}
}
}
}
\source{
TBD
}
\usage{
Cornejo2025
}
\description{
The Cornejo2025 longitudinal dataset as three-dimensional arrays,
with subjects in mode 1, features in mode 2 and time in mode 3.
}
\keyword{datasets}
