% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Jakobsen2025}
\alias{Jakobsen2025}
\title{Jakobsen2025 longitudinal dataset measured mother-infant dyads}
\format{
\subsection{'Cornejo2025'}{

A list object with three elements:
\describe{
\item{faeces}{List object of the infant faecal longitudinal microbiota data.}
\item{milkMicrobiome}{List object of the mother milk longitudinal microbiota data.}
\item{milkMetabolome}{List object of the mother milk longitudinal metabolomics data.}
}
}
}
\source{
TBD
}
\usage{
Jakobsen2025
}
\description{
The Jakobsen2025 longitudinal dataset as three-dimensional arrays,
with subjects in mode 1, features in mode 2 and time in mode 3.
}
\keyword{datasets}
