% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cornejo2025}
\alias{Cornejo2025}
\title{Cornejo2025 longitudinal dataset measured in transgender persons}
\format{
\subsection{'Cornejo2025'}{

A list object with seven elements:
\describe{
\item{Tongue_microbiome}{List object of the tongue longitudinal microbiota data.}
\item{Salivary_microbiome}{List object of the saliva longitudinal microbiota data.}
\item{Salivary_cytokines}{List object of the longitudinal salivary cytokine data.}
\item{Salivary_biochemistry}{List object of the longitudinal salivary biochemistry data.}
\item{Circulatory_hormones}{List object of the longitudinal circulatory hormone data.}
\item{Clinical_measurements}{List object of the longitudinal clinical outcome data.}
\item{Subject_metadata}{Matrix with subject metadata.}
}
}
}
\source{
TBD
}
\usage{
Cornejo2025
}
\description{
The Cornejo2025 longitudinal dataset as three-dimensional arrays,
with subjects in mode 1, features in mode 2 and time in mode 3.
}
\keyword{datasets}
