% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Jakobsen2025}
\alias{Jakobsen2025}
\title{Jakobsen2025 longitudinal dataset measured in mother-infant dyads}
\format{
\subsection{'Jakobsen2025'}{

A list object with seven elements:
\describe{
\item{faeces}{List object of the longitudinal infant faecal microbiota data.}
\item{milkMicrobiome}{List object of the longitudinal HM microbiota data.}
\item{milkMetabolomics}{List object of the longitudinal salivary cytokine data.}
}
}
}
\source{
TBD
}
\usage{
Jakobsen2025
}
\description{
The Jakobsen2025 longitudinal dataset as three-dimensional arrays,
with subjects in mode 1, features in mode 2 and time in mode 3.
}
\keyword{datasets}
